% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_synonyms}
\alias{taxon_synonyms}
\title{Synonyms for bacterial taxa}
\format{
A data frame with the following columns:
\describe{
\item{name}{The synonym}
\item{rank}{The rank of the taxon}
\item{correct_name}{The "correct" name for the taxon, as it is used in
other package data.}
}
}
\usage{
taxon_synonyms
}
\description{
Synonyms for bacterial taxa
}
\keyword{datasets}
