% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what_phenotype}
\alias{what_phenotype}
\title{Fetch the phenotype values for a given lineage and phenotype}
\usage{
what_phenotype(
  lineage,
  phenotype,
  db = whatbacteria::taxon_phenotypes,
  synonyms = whatbacteria::taxon_synonyms
)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{phenotype}{The name of the column in \code{db} that contains the
phenotype of interest}

\item{db}{A data frame with columns named "taxon", "rank", and the column
name specified in \code{phenotype}}

\item{synonyms}{A data frame of taxonomic synonyms with columns "name" and
"correct_name"}
}
\value{
A vector of assigned phenotype values
}
\description{
Fetch the phenotype values for a given lineage and phenotype
}
\details{
This function operates much like \code{antibiotic_susceptibility}, except
that it pulls phenotype values from the database instead of susceptibility
information. To subsequently determine susceptibility from phenotype, this
function uses the named vector provided in the \code{susceptibility}
argument.

As a reminder, the taxonomic ranks, in order from highest to lowest, are
Kingdom, Phylum, Class, Order, Family, Genus, and Species. The ranks in the
database must be capitalized, exactly as they are written here.
}
\examples{
what_phenotype(
  c("Bacteroidetes", "Firmicutes", "Firmicutes; Negativicutes"),
  "gram_stain")
}
